Bitmap: extend overwrite_channel() support

+++ include/mitsuba/core/bitmap.h
@@ -313,7 +313,7 @@ public:
     void set_metadata(const Properties &metadata) { m_metadata = metadata; }
 
     /// Overwrites the desired channel with the given value.
-    void overwrite_channel(size_t channel, float value);
+    void overwrite_channel(size_t channel, double value);
 
     /// Clear the bitmap to zero
     void clear();
